Extract storage installation for easy reuse
This extracts the storage installation bits for easy reuse, so that
other implementations of the resource metrics API don't have to go
through as much work.
